test: 통합 테스트 fixture 제거 및 서비스 예외 테스트 추가
